In process of implementing text rendering using the JOGL library's TextRenderer class, experimented with drawing text using lines, finding suitable way to represent the main camera with a slice of the viewing volume (frustum)

+++ src/main/java/org/cytoscape/paperwing/internal/rendering/ResetSceneProcedure.java
@@ -30,9 +30,14 @@ public class ResetSceneProcedure implements ReadOnlyGraphicsProcedure {
 		Vector3 target = camera.getTarget();
 		Vector3 up = camera.getUp();
 
-		glu.gluLookAt((float) position.x(), (float) position.y(), (float) position.z(), 
-				(float) target.x(), (float) target.y(), (float) target.z(), 
-				(float) up.x(), (float) up.y(), (float) up.z());
+//		System.out.println(position);
+//		System.out.println(target);
+//		System.out.println(up);
+		
+		glu.gluLookAt(
+				position.x(), position.y(), position.z(), 
+				target.x(), target.y(), target.z(), 
+				up.x(), up.y(), up.z());
 	}
 
 }
